# Require Julia 1.9 and remove ScalarWrapper