--- a/src/utils.jl
+++ b/src/utils.jl
@@ -35,14 +35,6 @@ else
     end
 end
 
-mutable struct ScalarWrapper{T}
-    x::T
-end
-
-@inline Base.getindex(s::ScalarWrapper) = s.x
-@inline Base.setindex!(s::ScalarWrapper, v) = s.x = v
-Base.copy(s::ScalarWrapper{T}) where {T} = ScalarWrapper{T}(copy(s.x))
-
 convert_to_orderedset(set::AbstractVector{T}) where T = OrderedSet{T}(set)
 convert_to_orderedset(set::AbstractSet{T}) where T = convert(OrderedSet{T}, set)