This patch sets the Julia compat to 1.9. This make it possible to use `const` struct fields in `mutable structs` and therefore the `ScalarWrapper` struct can be removed.

TODO: Benchmarks.