@fredrikekre 2024-07-01
I don't see any measurable impact from this PR:
```
# Single quadratic vector field
#
# PR:     4.450 ms (9 allocations: 1.34 KiB)
# master: 4.395 ms (9 allocations: 1.34 KiB)
# PR (all):     96.963 ms (9 allocations: 1.34 KiB)
# master (all): 99.174 ms (9 allocations: 1.34 KiB)


# Single linear scalar field
#
# PR:     260.800 μs (9 allocations: 768 bytes)
# master: 263.413 μs (9 allocations: 768 bytes)
# PR (all):     5.219 ms (9 allocations: 768 bytes)
# master (all): 5.302 ms (9 allocations: 768 bytes)


# Both fields
#
# PR:     5.440 ms (9 allocations: 1.42 KiB)
# master: 5.411 ms (9 allocations: 1.42 KiB)
# PR: (all)     117.858 ms (9 allocations: 1.42 KiB)
# master (all): 122.170 ms (9 allocations: 1.42 KiB)
```

<details><summary>Code</summary>

```julia
using Ferrite
using BenchmarkTools

n = 20
grid = generate_grid(Hexahedron, (n, n, n))

all = true

if all
    addfacetset!(grid, "dΩ", x -> true)
    sort!(grid.facetsets["dΩ"], by = x -> (x[1], x[2]))
else
    dΩ = union(values(grid.facetsets)...)
    sort!(dΩ, by = x -> (x[1], x[2]))
    addfacetset!(grid, "dΩ", dΩ)
end

ip_u = Lagrange{RefHexahedron, 2}()^3
ip_v = Lagrange{RefHexahedron, 1}()
fqr_u = FacetQuadratureRule{RefHexahedron}(2)
fqr_v = FacetQuadratureRule{RefHexahedron}(1)
facetvalues_u = FacetValues(fqr_u, ip_u)
facetvalues_v = FacetValues(fqr_v, ip_v)


# Single quadratic vector field
dh = DofHandler(grid)
add!(dh, :u, ip_u)
close!(dh)
f = zeros(ndofs(dh))
fe = zeros(ndofs_per_cell(dh))

function bench_me(dh, f, fe, facetvalues)
    fill!(f, 0)
    b = Vec{3}((1.0, 2.0, 3.0))
    for facet in FacetIterator(dh, getfacetset(grid, "dΩ"))
        reinit!(facetvalues, facet)
        for qp in 1:getnquadpoints(facetvalues)
            for i in 1:getnbasefunctions(facetvalues)
                δu = shape_value(facetvalues, qp, i)
                fe[i] += (δu ⋅ b) * getdetJdV(facetvalues, qp)
            end
        end
        for (i, I) in pairs(celldofs(facet))
            f[I] += fe[i]
        end
    end
end

@btime bench_me($dh, $f, $fe, $facetvalues_u)

# Single quadratic vector field
#
# PR:     4.450 ms (9 allocations: 1.34 KiB)
# master: 4.395 ms (9 allocations: 1.34 KiB)
# PR (all):     96.963 ms (9 allocations: 1.34 KiB)
# master (all): 99.174 ms (9 allocations: 1.34 KiB)


# Single linear scalar field
dh = DofHandler(grid)
add!(dh, :v, ip_v)
close!(dh)
f = zeros(ndofs(dh))
fe = zeros(ndofs_per_cell(dh))

function bench_me2(dh, f, fe, facetvalues)
    fill!(f, 0)
    b = 1.0
    for facet in FacetIterator(dh, getfacetset(grid, "dΩ"))
        reinit!(facetvalues, facet)
        for qp in 1:getnquadpoints(facetvalues)
            for i in 1:getnbasefunctions(facetvalues)
                δu = shape_value(facetvalues, qp, i)
                fe[i] += δu * b * getdetJdV(facetvalues, qp)
            end
        end
        for (i, I) in pairs(celldofs(facet))
            f[I] += fe[i]
        end
    end
end

@btime bench_me2($dh, $f, $fe, $facetvalues_v)

# Single linear scalar field
#
# PR:     260.800 μs (9 allocations: 768 bytes)
# master: 263.413 μs (9 allocations: 768 bytes)
# PR (all):     5.219 ms (9 allocations: 768 bytes)
# master (all): 5.302 ms (9 allocations: 768 bytes)


# Both fields
dh = DofHandler(grid)
add!(dh, :u, ip_u)
add!(dh, :v, ip_v)
close!(dh)
f = zeros(ndofs(dh))
fe = zeros(ndofs_per_cell(dh))

function bench_me3(dh, f, fe, facetvalues_u, facetvalues_v)
    fill!(f, 0)
    bu = Vec{3}((1.0, 2.0, 3.0))
    bv = 1.0
    range_u = dof_range(dh, :u)
    range_v = dof_range(dh, :v)
    for facet in FacetIterator(dh, getfacetset(grid, "dΩ"))
        # u
        reinit!(facetvalues_u, facet)
        for qp in 1:getnquadpoints(facetvalues_u)
            for (i, I) in pairs(range_u)
                δu = shape_value(facetvalues_u, qp, i)
                fe[I] += δu ⋅ bu * getdetJdV(facetvalues_u, qp)
            end
        end
        # v
        reinit!(facetvalues_v, facet)
        for qp in 1:getnquadpoints(facetvalues_v)
            for (i, I) in pairs(range_v)
                δu = shape_value(facetvalues_v, qp, i)
                fe[I] += δu * bv * getdetJdV(facetvalues_v, qp)
            end
        end
        for (i, I) in pairs(celldofs(facet))
            f[I] += fe[i]
        end
    end
end

@btime bench_me3($dh, $f, $fe, $facetvalues_u, $facetvalues_v)

# Both fields
#
# PR:     5.440 ms (9 allocations: 1.42 KiB)
# master: 5.411 ms (9 allocations: 1.42 KiB)
# PR: (all)     117.858 ms (9 allocations: 1.42 KiB)
# master (all): 122.170 ms (9 allocations: 1.42 KiB)

```

</details>